feat: Install docker buildx and create builder instance
This commit adds the installation of docker buildx and the creation of a builder instance named "mybuilder". This is done to improve the build process for Docker images. The installation of docker buildx is necessary for cross-platform builds, and the creation of the builder instance allows for concurrent builds and improved performance.

Signed-off-by: Steffen Exler <steffen@exivity.com>